requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-extensions==4.1
django-timezone-field==7.1
djangorestframework==3.16.0
dnspython==2.7.0
eventlet==0.40.0
greenlet==3.2.2
idna==3.10
kombu==5.5.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
